WIP Refactor core IO, metadata, and layer helpers
Restructure IO and metadata responsibilities: move image/video readers, HDF readers, config read/write, and lazy image helpers into core/io; extract keypoint metadata population into new core/layers.populate_metadata; centralize provenance attachment as attach_source_and_io and use discovery/provenance helpers for kind/io construction. Update imports in _reader, _widgets and _writer to use the new core modules, and adjust pydantic model dumping/merging to use model_dump(mode="python") for safer serialization. This cleans up duplication, improves separation of concerns, and makes provenance/metadata handling deterministic and reusable.
