+++ src/napari_deeplabcut/core/io.py
@@ -1,185 +1,76 @@
-"""Deterministic discovery of DeepLabCut annotation artifacts in a folder.
-
-This module is pure logic (no napari imports). It does not change behavior until
-reader/writer code starts using it.
-
-Why this exists
----------------
-DeepLabCut folders often contain multiple H5/CSV artifacts side-by-side
-(e.g. multiple CollectedData_*.h5, and machinelabels*.h5). Any "first match wins"
-logic is unsafe and non-deterministic across OS/filesystems.
-
-This discovery layer:
-- Enumerates all relevant files.
-- Classifies them (GT vs machine).
-- Returns results in a deterministic order.
-
-Conventions supported
----------------------
-- Ground-truth: CollectedData*.h5 / CollectedData*.csv
-- Machine:      machinelabels*.h5 / machinelabels*.csv
-
-CSV-only folders are supported (e.g. shared data without H5). We treat those as
-valid sources, with an optional inferred H5 companion if present.
 """
+Core IO utilities.
 
+This module is intentionally small:
+- re-exports deterministic discovery helpers (for backward-compatible imports)
+- provides config write helpers
+
+Reader/writer and napari LayerData logic must live outside core/ for separation of concerns.
+"""
 # src/napari_deeplabcut/core/io.py
+
 from __future__ import annotations
 
-from collections.abc import Iterable
-from dataclasses import dataclass
+import json
+import logging
+from collections.abc import Any, Callable
 from pathlib import Path
 
+import cv2
+import dask.array as da
 import numpy as np
 import pandas as pd
 import yaml
+from dask import delayed
+from dask_image.imread import imread
+from napari.types import LayerData
 from natsort import natsorted
 
+from napari_deeplabcut import misc
 from napari_deeplabcut.config.models import AnnotationKind
+from napari_deeplabcut.core.layers import populate_keypoint_layer_metadata
+from napari_deeplabcut.core.metadata import attach_source_and_io
+from napari_deeplabcut.core.paths import canonicalize_path
 
-# -----------------------------------------------------------------------------
-# Data structures
-# -----------------------------------------------------------------------------
+logger = logging.getLogger(__name__)
 
-
-@dataclass(frozen=True)
-class AnnotationArtifact:
-    """A discovered annotation artifact (H5/CSV) with inferred semantics.
-
-    Attributes
-    ----------
-    kind:
-        Annotation kind (gt or machine), if identifiable by filename.
-    h5_path:
-        Path to the H5 file, if present.
-    csv_path:
-        Path to the CSV file, if present.
-    stem:
-        Filename stem used for pairing (e.g. CollectedData_John).
-    """
-
-    kind: AnnotationKind | None
-    h5_path: Path | None
-    csv_path: Path | None
-    stem: str
-
-    @property
-    def primary_path(self) -> Path | None:
-        """Return a preferred path for opening (H5 preferred over CSV)."""
-        return self.h5_path or self.csv_path
+# FIXME move to config/data_formats.py or similar if more formats are added
+SUPPORTED_IMAGES = (".jpg", ".jpeg", ".png")
+SUPPORTED_VIDEOS = (".mp4", ".mov", ".avi")
 
 
-# -----------------------------------------------------------------------------
-# Classification helpers
-# -----------------------------------------------------------------------------
+def _load_config(config_path: str):
+    with open(config_path) as file:
+        return yaml.safe_load(file)
 
 
-def _infer_kind_from_name(name: str) -> AnnotationKind | None:
-    """Infer artifact kind from filename."""
-    low = name.lower()
-    if low.startswith("collecteddata"):
-        return AnnotationKind.GT
-    if low.startswith("machinelabels"):
-        return AnnotationKind.MACHINE
-    return None
-
-
-def _is_relevant_artifact(p: Path) -> bool:
-    """Return True if a path looks like a DLC annotation artifact."""
-    if not p.is_file():
-        return False
-    low = p.name.lower()
-    if low.endswith(".h5") or low.endswith(".csv"):
-        return low.startswith("collecteddata") or low.startswith("machinelabels")
-    return False
-
-
-# -----------------------------------------------------------------------------
-# Public API
-# -----------------------------------------------------------------------------
-
-
-def discover_annotations(folder: str | Path) -> list[AnnotationArtifact]:
-    """Discover DLC annotation artifacts in a folder.
-
-    Parameters
-    ----------
-    folder:
-        Directory to scan.
-
-    Returns
-    -------
-    list[AnnotationArtifact]
-        Deterministically ordered list of artifacts. Ordering is stable across OS:
-        natsorted by primary filename.
-    """
-    root = Path(folder)
-    if not root.exists() or not root.is_dir():
-        return []
-
-    # Collect relevant H5 and CSV files
-    files = [p for p in root.iterdir() if _is_relevant_artifact(p)]
-
-    # Deterministic ordering by filename
-    files = natsorted(files, key=lambda p: p.name)
-
-    # Group by stem
-    by_stem: dict[str, dict[str, Path]] = {}
-    for p in files:
-        st = p.stem
-        entry = by_stem.setdefault(st, {})
-        if p.suffix.lower() == ".h5":
-            entry["h5"] = p
-        elif p.suffix.lower() == ".csv":
-            entry["csv"] = p
-
-    artifacts: list[AnnotationArtifact] = []
-    for stem in natsorted(by_stem.keys()):
-        entry = by_stem[stem]
-        h5 = entry.get("h5")
-        csv = entry.get("csv")
-        kind = _infer_kind_from_name(stem)
-        artifacts.append(AnnotationArtifact(kind=kind, h5_path=h5, csv_path=csv, stem=stem))
-
-    # Stable order again by primary filename (H5 preferred, else CSV)
-    artifacts = natsorted(
-        artifacts,
-        key=lambda a: (a.primary_path.name if a.primary_path else a.stem),
+# Read config file and create keypoint layer metadata
+def read_config(configname: str) -> list[LayerData]:
+    config = _load_config(configname)
+    # FIXME duplicated DLCHeader misc/models
+    header = misc.DLCHeader.from_config(config)
+    metadata = populate_keypoint_layer_metadata(
+        header,
+        size=config["dotsize"],
+        pcutoff=config["pcutoff"],
+        colormap=config["colormap"],
+        likelihood=np.array([1]),
     )
-    return artifacts
+    metadata["name"] = f"CollectedData_{config['scorer']}"
+    metadata["ndim"] = 3
+    metadata["property_choices"] = metadata.pop("properties")
+    metadata["metadata"]["project"] = str(Path(configname).parent)
+    conversion_tables = config.get("SuperAnimalConversionTables")
+    if conversion_tables is not None:
+        super_animal, table = conversion_tables.popitem()
+        metadata["metadata"]["tables"] = {super_animal: table}
+    return [(None, metadata, "points")]
 
 
-def discover_annotation_paths(folder: str | Path) -> list[Path]:
-    """Convenience wrapper: return primary paths to open (H5 preferred else CSV)."""
-    return [a.primary_path for a in discover_annotations(folder) if a.primary_path is not None]
-
-
-def iter_annotation_candidates(paths: Iterable[str | Path]) -> list[Path]:
-    """Given user-provided paths (files or folders), return annotation file candidates.
-
-    This is useful for future reader logic: a single entry point that expands folders.
-    The returned list is deterministic.
-    """
-    out: list[Path] = []
-    for p in paths:
-        pp = Path(p)
-        if pp.is_dir():
-            out.extend(discover_annotation_paths(pp))
-        elif _is_relevant_artifact(pp):
-            out.append(pp)
-
-    # Stable ordering
-    return natsorted(out, key=lambda x: x.name)
-
-
-# -----------------------------------------------------------------------------
-#  Write configuration utilities
-# -----------------------------------------------------------------------------
-
-
-def write_config(config_path: str, params: dict):
-    with open(config_path, "w") as file:
-        yaml.safe_dump(params, file)
+def write_config(config_path: str | Path, params: dict[str, Any]) -> None:
+    """Write DeepLabCut config.yaml parameters."""
+    with open(str(config_path), "w", encoding="utf-8") as f:
+        yaml.safe_dump(params, f)
 
 
 # -----------------------------------------------------------------------------
@@ -234,7 +125,7 @@ def read_hdf_single(file: Path, *, kind: AnnotationKind | None = None) -> list[L
     data[:, 0] = image_inds
     data[:, 1:] = df[["y", "x"]].to_numpy()
 
-    metadata = _populate_metadata(
+    metadata = populate_keypoint_layer_metadata(
         header,
         labels=df["bodyparts"],
         ids=df["individuals"],
@@ -250,11 +141,301 @@ def read_hdf_single(file: Path, *, kind: AnnotationKind | None = None) -> list[L
     if kind is not None:
         meta = metadata.setdefault("metadata", {})
         # Keep legacy source fields too
-        _attach_source_and_io(metadata, file)
+        attach_source_and_io(metadata, file)
         # Override kind in io to discovered kind
         if isinstance(meta.get("io"), dict):
             meta["io"]["kind"] = kind.value  # stored as enum value for JSON safety
     else:
-        _attach_source_and_io(metadata, file)
+        attach_source_and_io(metadata, file)
 
     return [(data, metadata, "points")]
+
+
+# ------------------------------------------------------------------------------
+# Superkeypoints diagram and JSON loading (for DLCHeader superkeypoints support)
+# -----------------------------------------------------------------------------
+def _load_superkeypoints_diagram(super_animal: str):
+    path = str(Path(__file__).parent / "assets" / f"{super_animal}.jpg")
+    try:
+        return imread(path), {"root": ""}, "images"
+    except Exception as e:
+        raise FileNotFoundError(f"Superkeypoints diagram not found for {super_animal}.") from e
+
+
+def _load_superkeypoints(super_animal: str):
+    path = str(Path(__file__).parent / "assets" / f"{super_animal}.json")
+    if not Path(path).is_file():
+        raise FileNotFoundError(f"Superkeypoints JSON file not found for {super_animal}.")
+    with open(path) as f:
+        return json.load(f)
+
+
+# Helper functions for lazy image reading and normalization
+# NOTE : forced keyword-only arguments for clarity
+def _read_and_normalize(*, filepath: Path, normalize_func: Callable[[np.ndarray], np.ndarray]) -> np.ndarray:
+    arr = cv2.imread(str(filepath), cv2.IMREAD_UNCHANGED)
+    if arr is None:
+        raise OSError(f"Could not read image: {filepath}")
+    return normalize_func(arr)
+
+
+def _normalize_to_rgb(arr: np.ndarray) -> np.ndarray:
+    if arr.ndim == 2:
+        return cv2.cvtColor(arr, cv2.COLOR_GRAY2RGB)
+    if arr.ndim == 3 and arr.shape[2] == 4:
+        return cv2.cvtColor(arr, cv2.COLOR_BGRA2RGB)
+    return cv2.cvtColor(arr, cv2.COLOR_BGR2RGB)
+
+
+def _expand_image_paths(path: str | Path | list[str | Path] | tuple[str | Path, ...]) -> list[Path]:
+    # Normalize input to list[Path]
+    raw_paths = [Path(p) for p in path] if isinstance(path, (list, tuple)) else [Path(path)]
+
+    expanded: list[Path] = []
+    for p in raw_paths:
+        if p.is_dir() and p.suffix.lower() != ".zarr":
+            file_matches: list[Path] = []
+            for ext in SUPPORTED_IMAGES:
+                file_matches.extend(p.glob(f"*{ext}"))
+            expanded.extend(x for x in natsorted(file_matches, key=str) if x.is_file())
+        else:
+            matches = list(p.parent.glob(p.name))
+            expanded.extend(matches or [p])
+
+    return [p for p in expanded if p.is_file() and p.suffix.lower() in SUPPORTED_IMAGES]
+
+
+# Lazy image reader that supports directories and lists of files
+def _lazy_imread(
+    filenames: str | Path | list[str | Path],
+    use_dask: bool | None = None,
+    stack: bool = True,
+) -> np.ndarray | da.Array | list[np.ndarray | da.Array]:
+    """Lazily reads one or more images with optional Dask support.
+
+    Resolves file paths using `_expand_image_paths`, ensuring consistent
+    handling of directories, glob patterns, and lists/tuples of paths.
+    Images are normalized to RGB and may be wrapped in Dask delayed
+    objects for lazy loading.
+
+    Behavior:
+        * If a single image is resolved:
+            - The image is read eagerly and returned as a NumPy array.
+        * If multiple images are resolved:
+            - The first image is read eagerly to determine shape and dtype.
+            - Subsequent images are loaded lazily via Dask unless
+              `use_dask=False`.
+            - Stacking behavior is controlled by `stack`.
+
+    Args:
+        filenames (str | Path | list[str | Path]):
+            File path(s), directory, or glob pattern(s) to load.
+        use_dask (bool | None, optional):
+            Whether to load images lazily using Dask.
+            Defaults to `True` when multiple files are found, otherwise
+            `False`.
+        stack (bool, optional):
+            If True, stack images along axis 0 into a single array.
+            If False, return a list of arrays or delayed arrays.
+            Defaults to True.
+
+    Returns:
+        np.ndarray | da.Array | list[np.ndarray | da.Array]:
+            Loaded image data. The return type depends on the number of
+            images found, the `use_dask` flag, and the `stack` option.
+
+    Raises:
+        ValueError: If no supported images are found.
+    """
+    expanded = _expand_image_paths(filenames)
+
+    if not expanded:
+        raise ValueError(f"No supported images were found for input: {filenames}")
+
+    if use_dask is None:
+        use_dask = len(expanded) > 1
+
+    images = []
+    first_shape = None
+    first_dtype = None
+
+    def make_delayed_array(fp: Path, first_shape: tuple[int, ...], first_dtype: np.dtype) -> da.Array:
+        """Create a dask array for a single file."""
+        return da.from_delayed(
+            delayed(_read_and_normalize)(filepath=fp, normalize_func=_normalize_to_rgb),
+            shape=first_shape,
+            dtype=first_dtype,
+        )
+
+    for fp in expanded:
+        if first_shape is None:
+            arr0 = _read_and_normalize(filepath=fp, normalize_func=_normalize_to_rgb)
+            first_shape = arr0.shape
+            first_dtype = arr0.dtype
+
+            if use_dask:
+                images.append(make_delayed_array(fp, first_shape, first_dtype))
+            else:
+                images.append(arr0)
+            continue
+
+        if use_dask:
+            images.append(make_delayed_array(fp, first_shape, first_dtype))
+        else:
+            images.append(_read_and_normalize(filepath=fp, normalize_func=_normalize_to_rgb))
+
+    if len(images) == 1:
+        return images[0]
+
+    try:
+        return da.stack(images) if use_dask and stack else (np.stack(images) if stack else images)
+    except ValueError as e:
+        raise ValueError(
+            "Cannot stack images with different shapes using NumPy. "
+            "Ensure all images have the same shape or set stack=False."
+        ) from e
+
+
+# Read images from a list of files or a glob/string path
+def read_images(path: str | Path | list[str | Path]):
+    """Reads one or multiple images and returns a Napari Image layer.
+
+    Uses `_expand_image_paths` to resolve the input into a list of valid
+    image files. Supports single paths, glob expressions, directories,
+    and lists or tuples of such paths.
+
+    Behavior:
+        * If one file is found:
+            - Loaded using `dask_image.imread.imread`.
+        * If multiple files are found:
+            - Loaded lazily using `lazy_imread` into a stacked image
+              layer.
+
+    Args:
+        path (str | Path | list[str | Path]):
+            Input path(s), directory, or glob pattern(s) to expand into
+            supported image files.
+
+    Returns:
+        list[LayerData]:
+            A list containing one Napari layer tuple of the form
+            `(data, metadata, "image")`.
+
+    Raises:
+        OSError: If no supported images are found after expansion.
+    """
+    filepaths = _expand_image_paths(path)
+
+    if not filepaths:
+        raise OSError(f"No supported images were found in {path}")
+
+    filepaths = natsorted(filepaths, key=str)
+
+    # Multiple images → lazy-imread stack
+    if len(filepaths) > 1:
+        relative_paths = [canonicalize_path(fp, 3) for fp in filepaths]
+        params = {
+            "name": "images",
+            "metadata": {
+                "paths": relative_paths,
+                "root": str(filepaths[0].parent),
+            },
+        }
+        data = _lazy_imread(filepaths, use_dask=True, stack=True)
+        return [(data, params, "image")]
+
+    # Single image → old behavior
+    image_path = filepaths[0]
+    params = {
+        "name": "images",
+        "metadata": {
+            "paths": [canonicalize_path(image_path, 3)],
+            "root": str(image_path.parent),
+        },
+    }
+    return [(imread(str(image_path)), params, "image")]
+
+
+# Read HDF file and create keypoint layers
+def read_hdf(filename: str) -> list[LayerData]:
+    layers = []
+    for file in Path(filename).parent.glob(Path(filename).name):
+        layers.extend(read_hdf_single(file))
+    return layers
+
+
+# Video reader using OpenCV
+class Video:
+    def __init__(self, video_path):
+        if not Path(video_path).is_file():
+            raise ValueError(f'Video path "{video_path}" does not point to a file.')
+
+        self.path = video_path
+        self.stream = cv2.VideoCapture(video_path)
+        if not self.stream.isOpened():
+            raise OSError("Video could not be opened.")
+
+        self._n_frames = int(self.stream.get(cv2.CAP_PROP_FRAME_COUNT))
+        self._width = int(self.stream.get(cv2.CAP_PROP_FRAME_WIDTH))
+        self._height = int(self.stream.get(cv2.CAP_PROP_FRAME_HEIGHT))
+        self._frame = cv2.UMat(self._height, self._width, cv2.CV_8UC3)
+
+    def __len__(self):
+        return self._n_frames
+
+    @property
+    def width(self):
+        return self._width
+
+    @property
+    def height(self):
+        return self._height
+
+    def set_to_frame(self, ind):
+        ind = min(ind, len(self) - 1)
+        ind += 1  # Unclear why this is needed at all
+        self.stream.set(cv2.CAP_PROP_POS_FRAMES, ind)
+
+    def read_frame(self):
+        self.stream.retrieve(self._frame)
+        cv2.cvtColor(self._frame, cv2.COLOR_BGR2RGB, self._frame, 3)
+        return self._frame.get()
+
+    def close(self):
+        self.stream.release()
+
+
+def read_video(filename: str, opencv: bool = True):
+    if opencv:
+        stream = Video(filename)
+        # NOTE construct output shape tuple in (H, W, C) order to match read_frame() data
+        shape = stream.height, stream.width, 3
+
+        def _read_frame(ind):
+            stream.set_to_frame(ind)
+            return stream.read_frame()
+
+        lazy_reader = delayed(_read_frame)
+    else:  # pragma: no cover
+        from pims import PyAVReaderIndexed
+
+        try:
+            stream = PyAVReaderIndexed(filename)
+        except ImportError:
+            raise ImportError("`pip install av` to use the PyAV video reader.") from None
+
+        shape = stream.frame_shape
+        lazy_reader = delayed(stream.get_frame)
+
+    movie = da.stack([da.from_delayed(lazy_reader(i), shape=shape, dtype=np.uint8) for i in range(len(stream))])
+    elems = list(Path(filename).parts)
+    elems[-2] = "labeled-data"
+    elems[-1] = Path(elems[-1]).stem  # + Path(filename).suffix
+    root = str(Path(*elems))
+    params = {
+        "name": filename,
+        "metadata": {
+            "root": root,
+        },
+    }
+    return [(movie, params, "image")]
